redid addind <li> with needed class and input functions to data.js
